# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinTrack API — investment portfolio management and project task tracking
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portfolios/{portfolioId}/transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions in a portfolio
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Log a transaction
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Projects ──────────────────────────────────────────────────────────────
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with its tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Tasks ─────────────────────────────────────────────────────────────────
  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Members ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, portfolioId, ticker, type, quantity, price, total, transactedAt, createdAt]
      properties:
        id:
          type: integer
        portfolioId:
          type: integer
        ticker:
          type: string
        type:
          type: string
          enum: [buy, sell, dividend, deposit, withdrawal]
        quantity:
          type: number
        price:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        transactedAt:
          type: string
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [ticker, type, quantity, price, transactedAt]
      properties:
        ticker:
          type: string
          minLength: 1
        type:
          type: string
          enum: [buy, sell, dividend, deposit, withdrawal]
        quantity:
          type: number
        price:
          type: number
        notes:
          type: string
        transactedAt:
          type: string

    # ── Project schemas ─────────────────────────────────────────────────────
    Project:
      type: object
      required: [id, name, status, progress, taskCount, completedCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, on_hold, completed, archived]
        color:
          type: ["string", "null"]
        progress:
          type: integer
        taskCount:
          type: integer
        completedCount:
          type: integer
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [active, on_hold, completed, archived]
        color:
          type: string
        dueDate:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [active, on_hold, completed, archived]
        color:
          type: string
        dueDate:
          type: string

    # ── Task schemas ────────────────────────────────────────────────────────
    Task:
      type: object
      required: [id, projectId, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        ownerInitials:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        ownerId:
          type: integer
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        ownerId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]

    # ── Member schemas ──────────────────────────────────────────────────────
...