# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FinTrack API — investment portfolio management and project task tracking
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';
import type { TaskInputStatus } from './taskInputStatus';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status: TaskInputStatus;
  priority: TaskInputPriority;
  ownerId?: number;
  dueDate?: string;
}